CIV-7117 Refactor helper class
