+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/camunda/notification/NotifyHearingNoticeDefendantHandler.java
@@ -18,7 +18,7 @@ import java.util.Map;
 
 import static uk.gov.hmcts.reform.civil.callback.CallbackType.ABOUT_TO_SUBMIT;
 import static uk.gov.hmcts.reform.civil.callback.CaseEvent.NOTIFY_HEARING_NOTICE_DEFENDANT;
-import static uk.gov.hmcts.reform.civil.utils.JudicialDecisionNotificationUtil.isWithoutNotice;
+import static uk.gov.hmcts.reform.civil.utils.JudicialDecisionNotificationUtil.isWithNotice;
 
 @Service
 @RequiredArgsConstructor
@@ -46,7 +46,7 @@ public class NotifyHearingNoticeDefendantHandler extends CallbackHandler {
 
         CaseData caseData = callbackParams.getCaseData();
         try {
-            if (!isWithoutNotice(caseData)) {
+            if (isWithNotice(caseData)) {
                 caseData = hearingScheduledNotificationService.sendNotificationForDefendant(caseData);
             }
         } catch (NotificationException notificationException) {
