+++ src/test/java/uk/gov/hmcts/reform/civil/service/JudicialApplicantNotificationServiceTest.java
@@ -30,6 +30,7 @@ import uk.gov.hmcts.reform.civil.model.genapplication.GAJudicialWrittenRepresent
 import uk.gov.hmcts.reform.civil.model.genapplication.GAPbaDetails;
 import uk.gov.hmcts.reform.civil.model.genapplication.GARespondentOrderAgreement;
 import uk.gov.hmcts.reform.civil.model.genapplication.GASolicitorDetailsGAspec;
+import uk.gov.hmcts.reform.civil.model.genapplication.GAUrgencyRequirement;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -50,6 +51,8 @@ import static uk.gov.hmcts.reform.civil.enums.dq.GAJudgeMakeAnOrderOption.GIVE_D
 import static uk.gov.hmcts.reform.civil.enums.dq.GAJudgeRequestMoreInfoOption.REQUEST_MORE_INFORMATION;
 import static uk.gov.hmcts.reform.civil.enums.dq.GAJudgeRequestMoreInfoOption.SEND_APP_TO_OTHER_PARTY;
 import static uk.gov.hmcts.reform.civil.utils.ElementUtils.element;
+import static uk.gov.hmcts.reform.civil.utils.ElementUtils.wrapElements;
+import static uk.gov.hmcts.reform.civil.utils.JudicialDecisionNotificationUtil.isNotificationCriteriaSatisfied;
 
 @SpringBootTest(classes = {
     JudicialNotificationService.class,
@@ -79,6 +82,8 @@ class JudicialApplicantNotificationServiceTest {
     private CoreCaseDataService coreCaseDataService;
 
     private static final String APPLICANT = "applicant";
+
+    private static final String RESPONDENT = "respondent@email.com";
 
     private static final Long CASE_REFERENCE = 111111L;
     private static final String DUMMY_EMAIL = "hmcts.civil@gmail.com";
@@ -1251,6 +1256,36 @@ class JudicialApplicantNotificationServiceTest {
                 "general-apps-judicial-notification-make-decision-" + CASE_REFERENCE
             );
         }
+    }
+
+    @Test
+    void testIsNotificationCriteriaSatisfied() {
+        assertThat(isNotificationCriteriaSatisfied(getCaseData(NO, NO, NO, null))).isFalse();
+        assertThat(isNotificationCriteriaSatisfied(getCaseData(NO, NO, NO, RESPONDENT))).isFalse();
+        assertThat(isNotificationCriteriaSatisfied(getCaseData(NO, NO, YES, null))).isFalse();
+        assertThat(isNotificationCriteriaSatisfied(getCaseData(NO, NO, YES, RESPONDENT))).isTrue();
+        assertThat(isNotificationCriteriaSatisfied(getCaseData(NO, YES, NO, null))).isFalse();
+        assertThat(isNotificationCriteriaSatisfied(getCaseData(NO, YES, NO, RESPONDENT))).isFalse();
+        assertThat(isNotificationCriteriaSatisfied(getCaseData(NO, YES, YES, null))).isFalse();
+        assertThat(isNotificationCriteriaSatisfied(getCaseData(NO, YES, YES, RESPONDENT))).isFalse();
+        assertThat(isNotificationCriteriaSatisfied(getCaseData(YES, NO, NO, null))).isFalse();
+        assertThat(isNotificationCriteriaSatisfied(getCaseData(YES, NO, NO, RESPONDENT))).isFalse();
+        assertThat(isNotificationCriteriaSatisfied(getCaseData(YES, NO, YES,  null))).isFalse();
+        assertThat(isNotificationCriteriaSatisfied(getCaseData(YES, NO, YES,  RESPONDENT))).isFalse();
+        assertThat(isNotificationCriteriaSatisfied(getCaseData(YES, YES, NO, null))).isFalse();
+        assertThat(isNotificationCriteriaSatisfied(getCaseData(YES, YES, NO, RESPONDENT))).isFalse();
+        assertThat(isNotificationCriteriaSatisfied(getCaseData(YES, YES, YES, null))).isFalse();
+        assertThat(isNotificationCriteriaSatisfied(getCaseData(YES, YES, YES, RESPONDENT))).isFalse();
+    }
+
+    private CaseData getCaseData(YesOrNo isConsented, YesOrNo isUrgent, YesOrNo informOtherParty, String recipient) {
+        return CaseData.builder()
+            .generalAppRespondentAgreement(GARespondentOrderAgreement.builder().hasAgreed(isConsented).build())
+            .generalAppUrgencyRequirement(GAUrgencyRequirement.builder().generalAppUrgency(isUrgent).build())
+            .generalAppInformOtherParty(GAInformOtherParty.builder().isWithNotice(informOtherParty).build())
+            .generalAppRespondentSolicitors(wrapElements(GASolicitorDetailsGAspec.builder()
+                                                             .email(recipient).build()))
+            .build();
     }
 
     private CaseData caseDataForJudicialRequestForInformationOfApplication(
